+++ Memory.cpp
@@ -2,11 +2,11 @@
 #include <cstdint>
 
 // returns value held at given memory address. 
-int16_t Memory::read(uint16_t addr) const {
+int8_t Memory::read(uint8_t addr) const {
   return ram[addr];
 }
 
-int16_t Memory::write(uint16_t addr, int16_t value) {
+int8_t Memory::write(uint8_t addr, int8_t value) {
   ram[addr] = value;
   return ram[addr];
 }
